src/requirements/common.txt
numpy==1.8.1
Django==1.6.11
PyYAML==3.10
South==1.0.2
amqp==1.3.3
anyjson==0.3.3
billiard==3.3.0.13
celery==3.1.7
djorm-pgarray==1.0.3
greenlet==0.4.2
Cython==0.20
git+git://github.com/surfly/gevent.git@1.0rc3#egg=gevent
django-auth-ldap==1.1.7
django-filer==0.9.5
django-extensions==1.2.5
django-passwords==0.2.0
django-redis-cache==0.10.2
django-redis-status==1.0.0
django-websocket-redis==0.2.3
easy-thumbnails==1.5
hiredis==0.2.0
kombu==3.0.8
psycopg2==2.5.1
python-ldap==2.4.13
pytz==2013.8
six==1.4.1
sqlparse==0.1.10
ipython==1.2.0
mock==1.0.1
pandas==0.14.0
raven==5.1.1
django-tags-input==2.0.0
django-mptt==0.7.4
django-polymorphic==0.7.1
Pillow==2.8.1
python-dateutil==2.4.2
redis==2.10.3
uWSGI==2.0.10
cmsplugin-filer==0.10.1
django-appconf==1.0.1
django-classy-tags==0.6.1
django-cms==3.1.0
django-reversion==1.8.7
django-sekizai==0.8.1
Django-Select2==4.3.1
django-treebeard==3.0
djangocms-admin-style==0.2.7
djangocms-file==0.1
djangocms-flash==0.2.0
djangocms-googlemap==0.2
djangocms-inherit==0.1
djangocms-link==1.6.1
djangocms-picture==0.1
djangocms-snippet==1.5
djangocms-teaser==0.1
djangocms-text-ckeditor==2.5.1
djangocms-video==0.1
html5lib==0.99999
djangocms-column==1.5
djangocms-installer==0.7.6
djangocms-style==1.5
requests==2.7.0
tzlocal==1.1.3
Unidecode==0.4.17
